Move warning to the AidSelectionStep, fix React component
